Image mods

+++ meltdown/utils.py
@@ -603,7 +603,7 @@ class Utils:
 
         return key, value
 
-    def time_in(self, text: str, time_start: int, time_end: int) -> str:
+    def time_in(self, text: str, time_start: float, time_end: float) -> str:
         num = int(time_end - time_start)
         secs = utils.singular_or_plural(num, "sec", "secs")
         return f"Image generated in {num} {secs}"
